test(e2e): extract wheel fault phases
Keep the rear-left wheel-fault scenario as the top-level happy path while moving setup, artifact loading, cleanup, and staged report assertions into reusable Docker E2E helpers.

Closes #2857
